requirements.txt
# Core dependencies
Flask==3.0.3
Flask-CORS==6.0.1
requests==2.32.3

# Database
psycopg2-binary==2.9.11

# Authentication
bcrypt==3.2.0
PyJWT==2.8.0

# Environment variables
python-dotenv==0.21.0
